# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VLTONE Platform API
  version: 1.0.0
  description: Source-of-truth contract shared by the public site, admin site, and desktop application.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/browser-backgrounds/{backgroundID}/image:
    get:
      operationId: getPublicBrowserBackgroundImage
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/browser-backgrounds/{backgroundID}/thumbnail:
    get:
      operationId: getPublicBrowserBackgroundThumbnail
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/admin/browser-backgrounds:
    get:
      operationId: listAdminBrowserBackgrounds
      security: [{ adminCookie: [] }]
      responses:
        "200": { description: Backgrounds ordered by sort_order then newest first; public list includes only published images, content: { application/json: { schema: { $ref: "#/components/schemas/BrowserBackgroundList" } } } }
    post:
      operationId: uploadBrowserBackground
      security: [{ adminCookie: [], csrf: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, file]
              properties:
                title: { type: string, minLength: 1, maxLength: 160 }
                file: { type: string, format: binary, description: PNG/JPEG/WebP up to 10 MiB and 16 MP; server re-encodes and generates a thumbnail }
      responses:
        "201": { description: Published background, content: { application/json: { schema: { $ref: "#/components/schemas/BrowserBackground" } } } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/admin/browser-backgrounds/{backgroundID}:
    parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
    put:
      operationId: updateBrowserBackground
      security: [{ adminCookie: [], csrf: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, published, sort_order]
              properties:
                title: { type: string, minLength: 1, maxLength: 160 }
                published: { type: boolean }
                sort_order: { type: integer, minimum: 0, maximum: 100000 }
      responses:
        "200": { description: Updated metadata, content: { application/json: { schema: { $ref: "#/components/schemas/BrowserBackground" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteBrowserBackground
      security: [{ adminCookie: [], csrf: [] }]
      responses:
        "204": { description: Removed from collection and server storage; existing client copies are retained }
        "404": { $ref: "#/components/responses/Error" }
  /v1/admin/browser-backgrounds/{backgroundID}/image:
    get:
      operationId: getAdminBrowserBackgroundImage
      security: [{ adminCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/admin/browser-backgrounds/{backgroundID}/thumbnail:
    get:
      operationId: getAdminBrowserBackgroundThumbnail
      security: [{ adminCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /v1/releases:
    get:
      operationId: listReleases
      parameters: [{ $ref: "#/components/parameters/Locale" }]
      responses:
        "200":
          description: Published releases ordered by semantic version
          content:
            application/json:
              schema:
                type: object
                required: [releases]
                properties: { releases: { type: array, items: { $ref: "#/components/schemas/PublicRelease" } } }
  /v1/releases/latest:
    get:
      operationId: getLatestRelease
      parameters: [{ $ref: "#/components/parameters/Platform" }, { $ref: "#/components/parameters/Locale" }]
      responses:
        "200": { description: Latest release with an installer for the requested platform, content: { application/json: { schema: { $ref: "#/components/schemas/LatestRelease" } } } }
        "204": { description: No published release supports this platform }
  /v1/releases/{version}:
    get:
      operationId: getRelease
      parameters: [{ $ref: "#/components/parameters/ReleaseVersion" }, { $ref: "#/components/parameters/Locale" }]
      responses:
        "200": { description: Localized published release, content: { application/json: { schema: { $ref: "#/components/schemas/PublicRelease" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/releases/{version}/download/{kind}:
    get:
      operationId: downloadReleaseArtifact
      parameters: [{ $ref: "#/components/parameters/ReleaseVersion" }, { $ref: "#/components/parameters/ArtifactKind" }]
      responses:
        "200": { description: Installer with Range, ETag and SHA-256 headers, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/releases/{version}/screenshots/{screenshotId}:
    get:
      operationId: getReleaseScreenshot
      parameters: [{ $ref: "#/components/parameters/ReleaseVersion" }, { $ref: "#/components/parameters/ScreenshotId" }]
      responses:
        "200": { description: Sanitized release screenshot, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/web/auth/register:
    post:
      operationId: register
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterRequest" } } }
      responses:
        "201": { $ref: "#/components/responses/AccountSession" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/web/auth/login:
    post:
      operationId: webLogin
      requestBody: { $ref: "#/components/requestBodies/Login" }
      responses:
        "200": { $ref: "#/components/responses/AccountSession" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/web/auth/logout:
    post:
      operationId: webLogout
      security: [{ webCookie: [], csrf: [] }]
      responses: { "204": { description: Signed out }, "403": { $ref: "#/components/responses/Error" } }
  /v1/web/auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [email], properties: { email: { type: string, format: email } } } } }
      responses: { "202": { description: Accepted regardless of account existence } }
  /v1/web/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PasswordResetConfirmRequest" } } }
      responses: { "204": { description: Password changed and sessions revoked }, "400": { $ref: "#/components/responses/Error" } }
  /v1/me:
    get:
      operationId: getMe
      security: [{ webCookie: [] }]
      responses: { "200": { $ref: "#/components/responses/AccountSession" } }
  /v1/me/quota:
    get:
      operationId: getMyQuota
      security: [{ webCookie: [] }]
      responses: { "200": { description: Current UTC cycle, content: { application/json: { schema: { $ref: "#/components/schemas/Quota" } } } } }
  /v1/me/devices:
    get:
      operationId: listMyDevices
      security: [{ webCookie: [] }]
      responses:
        "200":
          description: Devices, including revoked history and the active-device limit
          content:
            application/json:
              schema:
                type: object
                required: [devices, limit]
                properties:
                  devices: { type: array, items: { $ref: "#/components/schemas/Device" } }
                  limit: { const: 2 }
  /v1/me/devices/{deviceId}:
    delete:
      operationId: revokeMyDevice
      security: [{ webCookie: [], csrf: [] }]
      parameters: [{ $ref: "#/components/parameters/DeviceId" }]
      responses: { "204": { description: Revoked }, "404": { $ref: "#/components/responses/Error" } }
  /v1/bug-reports:
    post:
      operationId: createBugReport
      security: [{ webCookie: [], csrf: [] }]
      requestBody:
        required: true
        content: { multipart/form-data: { schema: { $ref: "#/components/schemas/BugReportRequest" } } }
      responses: { "201": { description: One-way report receipt }, "422": { $ref: "#/components/responses/Error" } }
  /v1/desktop/auth/login:
    post:
      operationId: desktopLogin
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DesktopLoginRequest" } } }
      responses: { "200": { $ref: "#/components/responses/DesktopSession" }, "409": { $ref: "#/components/responses/Error" } }
  /v1/desktop/auth/refresh:
    post:
      operationId: desktopRefresh
      parameters: [{ name: Idempotency-Key, in: header, required: false, schema: { type: string, format: uuid }, description: Durable refresh operation ID; resend the same ID after a lost response }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DesktopRefreshRequest" } } }
      responses: { "200": { $ref: "#/components/responses/DesktopSession" }, "401": { $ref: "#/components/responses/Error" } }
  /v1/desktop/auth/logout:
    post:
      operationId: desktopLogout
      security: [{ desktopBearer: [] }]
      responses: { "204": { description: Revoked } }
  /v1/desktop/me:
    get:
      operationId: desktopMe
      security: [{ desktopBearer: [] }]
      responses: { "200": { $ref: "#/components/responses/AccountSession" } }
...
  /v1/desktop/telemetry/batch:
    post:
      operationId: uploadTelemetryBatch
      security: [{ reporterBearer: [] }, { desktopBearer: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TelemetryBatch" } } }
      responses: { "202": { description: Idempotent batch accepted }, "422": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    webCookie: { type: apiKey, in: cookie, name: vlt_web_session }
    adminCookie: { type: apiKey, in: cookie, name: vlt_admin_session }
    csrf: { type: apiKey, in: header, name: X-CSRF-Token }
    desktopBearer: { type: http, scheme: bearer, bearerFormat: Ed25519 }
    reporterBearer: { type: http, scheme: bearer, bearerFormat: opaque }
  parameters:
    BackgroundID: { name: backgroundID, in: path, required: true, schema: { type: string, format: uuid } }
...
    ReleaseVersion: { name: version, in: path, required: true, schema: { type: string, pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$' } }
...
    Platform: { name: platform, in: query, required: true, schema: { type: string, enum: [windows, macos, linux] } }
    Locale: { name: locale, in: query, required: false, schema: { type: string, enum: [ru, en], default: en } }
...
  responses:
    Error:
      description: Structured API error
      content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } }
...
  schemas:
    BrowserBackgroundList:
      type: object
      required: [backgrounds]
      properties:
        backgrounds: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/BrowserBackground" } }
...
    LatestRelease:
      type: object
      required: [version, published_at, page_url]
      properties:
        version: { type: string }
        published_at: { type: string, format: date-time }
        page_url: { type: string, format: uri }
...